few changes on collisions

See merge request this-one-is-also-flying/2020_302_thisoneisalsoflying!141

+++ src/Domain/Settings.java
@@ -37,7 +37,7 @@ public class Settings {
 	private String username;
 	
 	private Dimension screenSize;
-	public double timeRemaining;
+	public int timeRemaining;
 	
 	public static int timeMult = 20;
